Rewrote sizeSelect to be more dry
